chore(conversation): move conversationparticipants fetch to ConversationResource
